requirements.txt
python-chess
stockfish
numpy
pandas
matplotlib